requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy
geoalchemy2
psycopg2-binary==2.9.9
python-dotenv
alembic
pydantic-settings
pytest
